requirements.txt
paho-mqtt==1.5.1
pymongo==3.12.0
